added IClientService interface to define client management operations and updated ClientService to implement this interface
